refactor clickhouse client, ditch typed generic functions and fixed field types conversions

+++ packages/shared/pkg/chdb/migrations/0001_create_metrics_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS metrics (
 	sandbox_id String,
 	team_id String,
 	cpu_count UInt32,
-	cpu_used_pct Float64,
+	cpu_used_pct Float32,
 	mem_total_mib UInt64,
 	mem_used_mib UInt64
 ) Engine MergeTree()
